Updated responses and added comments

+++ users.py
@@ -19,7 +19,7 @@ def authorization():
         else:
             return False
     else:
-        return False
+        return None
 
 
 class Users(Resource):
@@ -29,18 +29,22 @@ class Users(Resource):
         username = data['username']
         password = data['password']
         display_name = data['display_name']
+        # Check NULL condition of all fields
         if username == "" or password == "" or display_name == "":
             response = Response(status = 400, mimetype = 'application/json')
-            return jsonify(response)
+            return response
         email_validator = lepl.apps.rfc3696.Email()
         if not email_validator(username):
+            # Email validation
             response = Response(status = 400, mimetype = 'application/json')
-            return jsonify(response)
+            return response
         user_details = users_db.get_user_details(username)
         if user_details:
+            #User already exists
             response = Response(status = 409, mimetype = 'application/json')
             return response
         else:
+            # Create a new user
             users_db.create_user(username, password, display_name)
             response = Response(status = 201, mimetype = 'application/json')
             return response
@@ -50,42 +54,61 @@ class Users(Resource):
         username = data['username']
         password = data['password']
         new_password = data['new_password']
+        # Check NULL condition of all fields
         if username == "" or password == "" or new_password == "":
             response = Response(status = 400, mimetype = 'application/json')
-            return jsonify(response)
+            return response
         auth = authorization()
-        if auth:
-            authentication = request.authorization
-            if authentication.username == username and authentication.password == password:
-                users_db.update_password(username, new_password)
-                response = Response(status = 200, mimetype = 'application/json')
-                return response
+        # Check authentication
+        if auth is not None:
+            if auth:
+                authentication = request.authorization
+                if authentication.username == username and authentication.password == password:
+                    users_db.update_password(username, new_password)
+                    response = Response(status = 200, mimetype = 'application/json')
+                    return response
+                else:
+                    # Unauthorized access
+                    response = Response(status = 401, mimetype = 'application/json')
+                    return response
             else:
+                # Unauthorized access
                 response = Response(status = 401, mimetype = 'application/json')
                 return response
         else:
-            response = Response(status = 401, mimetype = 'application/json')
+            # User not found
+            response = Response(status = 404, mimetype = 'application/json')
             return response
 
     def delete(self):
         data = request.get_json()
         username = data['username']
         password = data['password']
+        # Check NULL condition of all fields
         if username == "" or password == "" or password == "":
             response = Response(status = 400, mimetype = 'application/json')
-            return jsonify(response)
+            return response
         auth = authorization()
-        if auth:
-            authentication = request.authorization
-            if authentication.username == data['username'] and authentication.password == data['password']:
-                users_db.delete_user(data)
-                response = Response(status = 200, mimetype = 'application/json')
-                return response
+        # Check authentication
+        if auth is not None:
+            if auth:
+                authentication = request.authorization
+                if authentication.username == data['username'] and authentication.password == data['password']:
+                    # Delete a user
+                    users_db.delete_user(data)
+                    response = Response(status = 200, mimetype = 'application/json')
+                    return response
+                else:
+                    # Unauthorized access
+                    response = Response(status = 401, mimetype = 'application/json')
+                    return response
             else:
+                # Unauthorized access
                 response = Response(status = 401, mimetype = 'application/json')
                 return response
         else:
-            response = Response(status = 401, mimetype = 'application/json')
+            # User not found
+            response = Response(status = 404, mimetype = 'application/json')
             return response
